chore: handle aggregation queries sequentially follow up (#8450)
This is a follow up on https://github.com/Unleash/unleash/pull/8446
- migrating some metrics (check the commits)
- moving event listener registration outside metric calculation (every 2
hs we are adding new listeners)
- adding jitter to metric refresh
- manually executing some of the tasks immediately to avoid changing the
test logic (with jitter now we'd have to force or wait until the first
execution of the scheduled task that collect the metrics). We might
decide to change this later after the migration
- try catch the execution of the task
- register metrics now "just registers" and returns functions for
refreshing metrics. In the future, we'll return an object with the
ability of independently updating specific metrics so we can tweak the
frequency of updates